requirements.txt
requests
sentence-transformers
faiss-cpu
numpy
pandas
openpyxl
colorama
spacy==3.7.2
tqdm
fastapi
uvicorn[standard]
pydantic
python-multipart
aiofiles
jinja2
openai
anthropic
pytest
pytest-asyncio
httpx